Switch to using Ref<Shortcut>
Switched from using Ref<InputEvent> to Ref<Shortcut> for storage.  An issue arose where when using override_builtin_shortcut(int, Ref<InputEvent>) because if the the override was from the editor ED_GET_SHORTCUT()->get_shortcut() then the editor needed to be restarted for the change to take effect. This is because it was storing the Ref to the InputEvent, not the ref to the Shortcut. Shortcut's keep the same memory address even when their shortcut InputEvent is changed. So now you can change the shortcut from EditorSettings and a restart will not be required.